requirements.txt
openai>=1.12.0
pytube>=15.0.0
moviepy>=1.0.3
python-dotenv>=1.0.0
tqdm>=4.66.1
requests>=2.31.0
yt-dlp>=2023.11.16
